+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Fixed
 - Story History bug when story actions where triggered
+- Ticket not being reassigned to current user when state change to "started"
 
 ## [1.4.2] - 2017-05-24
 ### Removed